chore: Fix compilation errors and warnings on macOS

+++ crates/agentfs-core/src/types.rs
@@ -220,49 +220,49 @@ impl Attributes {
 
         // Add file type
         if self.is_dir {
-            mode |= libc::S_IFDIR;
+            mode |= libc::S_IFDIR as u32;
         } else if self.is_symlink {
-            mode |= libc::S_IFLNK;
+            mode |= libc::S_IFLNK as u32;
         } else if let Some(kind) = &self.special_kind {
             mode |= match kind {
-                SpecialNodeKind::Fifo => libc::S_IFIFO,
-                SpecialNodeKind::CharDevice { .. } => libc::S_IFCHR,
-                SpecialNodeKind::BlockDevice { .. } => libc::S_IFBLK,
-                SpecialNodeKind::Socket => libc::S_IFSOCK,
+                SpecialNodeKind::Fifo => libc::S_IFIFO as u32,
+                SpecialNodeKind::CharDevice { .. } => libc::S_IFCHR as u32,
+                SpecialNodeKind::BlockDevice { .. } => libc::S_IFBLK as u32,
+                SpecialNodeKind::Socket => libc::S_IFSOCK as u32,
             };
         } else {
-            mode |= libc::S_IFREG;
+            mode |= libc::S_IFREG as u32;
         }
 
         // Add permissions
         if self.mode_user.read {
-            mode |= libc::S_IRUSR;
+            mode |= libc::S_IRUSR as u32;
         }
         if self.mode_user.write {
-            mode |= libc::S_IWUSR;
+            mode |= libc::S_IWUSR as u32;
         }
         if self.mode_user.exec {
-            mode |= libc::S_IXUSR;
+            mode |= libc::S_IXUSR as u32;
         }
 
         if self.mode_group.read {
-            mode |= libc::S_IRGRP;
+            mode |= libc::S_IRGRP as u32;
         }
         if self.mode_group.write {
-            mode |= libc::S_IWGRP;
+            mode |= libc::S_IWGRP as u32;
         }
         if self.mode_group.exec {
-            mode |= libc::S_IXGRP;
+            mode |= libc::S_IXGRP as u32;
         }
 
         if self.mode_other.read {
-            mode |= libc::S_IROTH;
+            mode |= libc::S_IROTH as u32;
         }
         if self.mode_other.write {
-            mode |= libc::S_IWOTH;
+            mode |= libc::S_IWOTH as u32;
         }
         if self.mode_other.exec {
-            mode |= libc::S_IXOTH;
+            mode |= libc::S_IXOTH as u32;
         }
 
         mode
